Refactor controllers and services

+++ src/controllers/bookshelves.controller.ts
@@ -1,12 +1,13 @@
 import { plainToInstance } from "class-transformer";
 import { Request, Response } from "express";
-import { GetBookshelfByIdDto, SearchQueryDto } from "@dtos";
+import { SearchQueryDto } from "@dtos";
 import bookshelvesService from "@services/bookshelves.service";
 import {
   IBookshelf,
   IBookshelfWithoutBooks,
   IBookshelfWithUser,
 } from "../interfaces/bookshelves.interface";
+import { isNaN } from "lodash";
 
 export const getAllBookshelves = async (
   req: Request,
@@ -24,11 +25,10 @@ export const getBookshelfById = async (
   req: Request,
   res: Response
 ): Promise<Response<IBookshelf>> => {
-  const { id } = req.params;
-  if (!id) return res.status(400).send("Invalid ID parameter");
+  const bookshelfId = parseInt(req.params.bookshelfId, 10);
+  if (isNaN(bookshelfId)) return res.status(400).send("Invalid ID parameter");
 
-  const data: GetBookshelfByIdDto = { id: +id };
-  const bookshelf = await bookshelvesService.getBookshelfById(data);
+  const bookshelf = await bookshelvesService.getBookshelfById(bookshelfId);
   if (!bookshelf) return res.status(400).send("Bookshelf Not Found");
 
   return res.send(bookshelf);
@@ -38,10 +38,10 @@ export const getBookshelvesByUserId = async (
   req: Request,
   res: Response
 ): Promise<Response<IBookshelf[]>> => {
-  const { id } = req.params;
-  if (!id) return res.status(404).send("User Not Found");
+  const bookshelfId = parseInt(req.params.bookshelfId, 10);
+  if (isNaN(bookshelfId)) return res.status(400).send("Invalid ID parameter");
 
-  const bookshelf = await bookshelvesService.getBookshelvesByUserId(+id);
+  const bookshelf = await bookshelvesService.getBookshelfById(bookshelfId);
 
   return res.send(bookshelf);
 };
@@ -61,9 +61,10 @@ export const addBookToBookshelf = async (
   req: Request,
   res: Response
 ): Promise<Response<IBookshelf>> => {
-  const { id } = req.params;
+  const bookshelfId = parseInt(req.params.bookshelfId, 10);
+  if (isNaN(bookshelfId)) return res.status(400).send("Invalid ID parameter");
+
   const { bookIds } = req.body;
-  const bookshelfId: GetBookshelfByIdDto = { id: +id };
 
   const updatedBookshelf = await bookshelvesService.addBookToBookshelf(
     bookshelfId,
@@ -77,9 +78,11 @@ export const removeBooksFromBookshelf = async (
   req: Request,
   res: Response
 ): Promise<Response<IBookshelf>> => {
-  const { id } = req.params;
-  const { bookIds } = req.body;
-  const bookshelfId: GetBookshelfByIdDto = { id: +id };
+  const bookshelfId = parseInt(req.params.bookshelfId, 10);
+  if (isNaN(bookshelfId)) return res.status(400).send("Invalid ID parameter");
+
+  const bookIds = req.body.bookIds;
+
   const updatedBookshelf = await bookshelvesService.removeBooksFromBookshelf(
     bookshelfId,
     bookIds
@@ -91,9 +94,10 @@ export const updateBookshelf = async (
   req: Request,
   res: Response
 ): Promise<Response<IBookshelf>> => {
-  const { id } = req.params;
+  const bookshelfId = parseInt(req.params.bookshelfId, 10);
+  if (isNaN(bookshelfId)) return res.status(400).send("Invalid ID parameter");
+
   const updatedData = req.body;
-  const bookshelfId: GetBookshelfByIdDto = { id: +id };
 
   const updatedBookshelf = await bookshelvesService.updateBookshelf(
     bookshelfId,
@@ -103,21 +107,19 @@ export const updateBookshelf = async (
   return res.send(updatedBookshelf);
 };
 
-export const deleteBookshelf = async (
-  req: Request,
-  res: Response
-): Promise<Response<IBookshelfWithoutBooks>> => {
-  const { id } = req.params;
-  const bookshelfId: GetBookshelfByIdDto = { id: +id };
+export const deleteBookshelf = async (req: Request, res: Response) => {
+  const bookshelfId = parseInt(req.params.bookshelfId, 10);
+  if (isNaN(bookshelfId)) return res.status(400).send("Invalid ID parameter");
+
   const deletedBookshelf =
     await bookshelvesService.deleteBookshelf(bookshelfId);
+
   return res.send(deletedBookshelf);
 };
 
 export default {
   getAllBookshelves,
   getBookshelfById,
-  getBookshelvesByUserId,
   createBookshelf,
   addBookToBookshelf,
   removeBooksFromBookshelf,
